Move all tag-specific styles to processTag.js.
This reduces duplicate styles found in both Tag.svelte as well as SRSConstructor.js. All tags are now styled with inline styles on the tag as determined by processTag.js. Previously only dynamic styles were inline, but that meant that anything not dynamic had to be present in both of the previously mentioned files. This makes special tags have a single source of styling. margin between cards is still in both places, but I think final spacing and layout should be up to the consumer.