In preparation for the new extension rewrite, track and send an array of phrases that were already active (added, not deleted) when saving a search. The extension will use this to show existing phrases in another toast so if you refresh the page or look at an old search you don't have to go dig through the phrase list on the site to delete the phrase.
Also remove the old add route code for the tampermonkey script. It's so old nobody should be using it anymore.

Switched to counting phrases for the user by using a count call instead of checking the length of the list. Should be faster and use less memory since we don't actually need to get the whole phrase list for a user when adding phrases, it was inefficient.