Rewrote the extension and added settings.
There's now a settings page popup that shows when clicking on the extension button. There are currently only two settings, the search save mode (always/prompt/never) so you can look things up without feeling burdened by needing to delete everything depending on how you study, and another setting for notification position (left/right) which controls which top corner of the page the toasts show up at. They were always on the left but I always felt it was awkward, it'll now default to the right but can be configured to the left too. A NEW "storage" PERMISSION WAS NEEDED TO SAVE SETTINGS.

The extension was previously written in Vue, for consistency (and my preference) I rewrote it in Svelte. Additionally I added some new components to facilitate the new settings (prompt, and the settings page itself), and now a toast will show for any existing phrases in a search so that the user doesn't have to search through the list if they refreshed the page or re-searched for something that's active.

Also toasts no longer have the purple border-left. That was to match the site's theme, but the site has long since been redesigned and no longer has that element. It looked pretty strange on error toasts too.