Migrate from TypeORM to Prisma.
Prisma seems a lot better maintained and the API is a lot nicer. Currently not rewriting the initial migrations, no database schema changes were needed so I'll cross that bridge when I need to change it someday. The schema.prisma file was generated with database introspection and tweaked a little bit (@@map to use pascal case model names, and corrected some relations).

Swapped from the deprecated node-sass to sass.

Removed the old 'stats' collection which wasn't really used anymore, the prometheus/grafana setup is better.

Removed the really old db.ts which pre-dates the migration to TypeORM and was long unused.