Added a few new features/settings.
You can now disable the purple bar animation that indicates how long until a toast closes, and you can also initially have all toasts hidden. These are a couple settings to reduce motion and be less distracting if the user thinks they won't need to use the toasts very often.

Also added another setting that shows a badge on the action button indicating how many phrases you have active. Can be useful if you don't want to keep the site open or you forget to export and notice later that you have phrases that need to be exported. It polls every two hours and watches for web requests (THIS REQUIRES A NEW PERMISSION) to update the count when the user hits an API endpoint that's likely going to add/remove phrases.

The settings no longer use magic strings, everything is based off of a 'settingNames' map for easier auto-complete, it's more likely to be accurate. Also need to check for undefined and not just a falsey value when using a setting default because now there are booleans, don't want to ignore a setting that's been turned off!

A small bug was fixed where after a toast expires the 'deleted' state on a phrase was lost and it'd appear as if it hadn't been deleted. Switched to only hiding toasts with a .hidden class instead of unmounting it so the internal state of the phrase StasherToastPhrase component remains.

The extensions page now has a heading that more similarly matches the website's header (background color and border in particular). It also has a link to the site if you click the heading now.