Updated to the new sheodox-ui version. This includes a color scheme update mostly. I also changed up the three actions on the list that would collapse to a single dropdown menu to a dropdown menu and a delete button. I added both "edit" and "copy" actions to that dropdown.
This also fixes some css variable names that are copied into the card template which was causing improper CSS values (wrong fonts etc) and causing anki to assume the separator was a comma and not a semicolon. I also made the other dictionary links a little easier to click.

Because cards can be edited the code that looks for new phrases now looks at the ID of a phrase not the phrase text itself to find "new" phrases so it doesn't think that something that was just edited is new.