Added the option to choose to build cards that include the context sentence on the front. This eventually should be on a per-card basis, but for now it asks before making any cards which you want.
In general this works by storing the range for the selected text and splitting it into three chunks, the highlighted text and everything before and after it. Highlighted text gets sheodox-ui cyan badge styling in a `mark` element.

This is used on the front of the cards too. If using a context sentence on the front of the card it'll use smaller text because it has to fit more on the front.