Typescript on front end and general refactoring
I changed a bunch of things here for consistency and code quality. Introduced prettier/eslint/svelte-check which runs on a pre-commit hook. I also rewrote the front end with Typescript. ESM is used on the server now too. A lot of files were moved around or renamed to follow more consistent conventions. Frontend stuff is built with Vite instead of Webpack now.

I removed https from the nginx and certbot from the prod setup, as that'll be handled by HAProxy on the host machine. Similarly with Prometheus and Grafana, metrics collecting is handled outside of scope of this repo.

I moved the extension source files but haven't yet done any rewrite on that. This is focusing mostly on the website part for now. Linting is disabled for that as there are a lot of unaddressed problems and I haven't yet switched build tools from webpack.

The dev environment now uses some self signed certificates as Google OAuth requires https.