Added support for AnkiConnect
AnkiConnect is an Anki addon made for Yomichan that allows interacting with Anki over http requests. This makes adding Anki cards and decks very easy. This is the recommended way to import cards as the manual method is very error prone. The AnkiConnect integration allows you to select a deck and add cards to that deck, or create a deck first. This saves users from having to tweak the Anki import settings themselves.

It will check if the decks can be listed, then polls every couple seconds for it. This determines both the available decks and if AnkiConnect is available.

I couldn't use a proper Content-Type header on the requests, with it on there it was failing CORS even on the call to request permission.