If the user deletes all phrases after exporting it will now automatically reset the export
When the user deletes all the phrases it hides everything and says "you're done", but then if you get new phrases it'll try and add them on top of the cards you just exported. If the user just deletes the used phrases there's a "start over" button but not when they delete all the phrases. Now if they use the delete all button it'll automatically reset the export so next time it's fresh.