Don't show an empty list of decks if AnkiConnect connection status is unknown
The connection status store starts out as null, because of the way the conditional was set up it was showing an empty deck list first if AnkiConnect couldn't be reached. This just shows nothing for a brief moment. I thought about putting a loading indicator by now but it seems fairly quick to figure out if AnkiConnect is available so it sounded more distracting to show an animated loading indicator for a few frames than just showing nothing.
 Please enter the commit message for your changes. Lines starting