More consistent font sizes on cards
On desktop for some reason the definition text was awkwardly big. Meanwhile the context sentence on the front of the card was pretty small. I put a font size on the `.card` so it has a more consistent base style (the default font size on `.card` must be bigger than the rest of the font sizes we use). And upped the font size on the front of context cards.

I also removed the font size on the context sentence on the card back for consistency, it looks good with the new base size.