Improved the card customization
The button to enter customization is now more obvious, it's right between the reading field and "Add Card" as a plain "Customize" button.

The card customization has been redone a bit and now uses tabs for all the information. The left column is now all the editable stuff, and the Metadata which was above the preview is now on a separate tab. There's also a new tab for notes which allow you to type anything you want directly above or below the definition area, depending on where the user wants them. Notes are in a text area with `white-space: pre-line` so they can type multiple lines worth of text if they want.