Moved the customize button to somewhere less prominent.
In between the Add Card button and the reading field was a little too in your face, you'd have to look past it to get to the button you actually want most of the time. It's not going to be used very often so it's not important to be so forward with it.