Anki-Connect import error handling
The code to export cards using Anki-Connect didn't have any safety checks. It assumed things would always succeed, because it always had succeeded for me. I added some error handling and if an `error` is returned from Anki-Connect it will now show in an error toast and not set the flag for cards being imported, so if the problem is on the Anki side the user can fix it and try again.

I also made the "Add to Deck" button disabled if a deck hasn't been selected yet.

Anki-Connect is now properly spelled with a dash as it is on the official project home page instead of AnkiConnect.