Added a custom "Context.Reviews v1" card model. #44
Expecting there to be an existing card model called "Basic" with "Front" and "Back" fields lead to problems. Those names are all localized in other languages which leads to cards silently not being created. To get around that before adding the notes it will check to see if a "Context.Reviews v1" card model exists, and if it doesn't it will create it.

The card model is mostly the same as the "Basic" model, except that the CSS has been extracted and now exists on the model itself. This has some extra benefits: 1. less duplicated data between cards, 2. night mode on AnkiDroid works! AnkiDroid's night mode will try and invert colors on cards unless the card model includes a `night-mode` class in its css. It unfortunately doesn't look at css in the Front field of the card, but it will look at the model so that's not an issue anymore because we're not using Basic anymore. The html still has a lot of duplication between cards, but that can be improved with a version 2 of the card model along with other features I want to add.

The code is set up so the SRSConstructor just has to return a different model name (e.g. "Context.Reviews v2") and it will create that if needed, so further changes to the model only needs to change SRSConstructor.

The CSS still needs to exist sometimes, so it will be included in all previews on the site, and when exporting the manual text file.