Made hover styles more consistent between the import source choices and card style choice
They're now using a hybrid of both styles. 2px gray border, blue when hovered with a slight blue box shadow. Just helps with consistency. I didn't add the same lighter gray background on hover on the card style choices as they have a set background color and I wanted it to be an accurate representation of what they'll actually look like in Anki.