Redesigned the card list side bar and made other things more clear
The card list sidebar was kind of awkward and it didn't show you the full list of cards you've created. I made it now show those, as well as the other stuff all in individual fieldsets which seems to be more the design I wanted here than just headings. Additionally the phrase progress bar now fills to 100% when on the last phrase instead of stopping short (then disappearing when on the Export page).

I made the next phrase button switch to show "Export [num] Cards" when at the end, so it's obvious you can just keep hitting next.

When requesting permissions from Anki-Connect it now will show a toast reminding the user to look at Anki, the confirmation dialog can open in the background and then it's not obvious what's going on.

The phrase label now reads more like plain English, "Phrase 10 of 11" instead of "Phrases Processed (10/11)".