Re-add existing extension code
I had moved the folder when doing the typescript rewrite but because of a gitignore rule that ignored `extension` the new location was ignored (previously the folder name was `extension-src`, and I think the build destination was `extension`). No new changes have been made to the files as far as I can remember, just realizing they were gone from the remote git repository and went to investigate.