Added support for getting a list of phrases in the extension format.
The extension, for historical reasons that I'm not yet interested in refactoring, requires `{phrase: string, phrase_id: string}` instead of `{phrase: string, id: string}`, and the list endpoint didn't return stuff in that format yet. I added similar 'versioned' handling of that endpoint as the other endpoints. This is used to check if phrases are added already when in prompt mode so it knows if it should show a phrase as saved already or not.