Extension redesign, manifest v3
The extension works very similarly, but conceptually instead of it being a series of toasts it's a single popup/panel that contains all phrases added. There's no in-page count of phrases besides those that have been added too. I never really paid attention to it and it's duplicated by the browser toolbar button badge and the document title on the site.

When in 'prompt' mode the extension will split the search text into phrases just like how the server would. That way if you accidentally select more than one sentence you can easily just not save the phrases you don't want, instead of saving all then deleting the extra ones. This makes 'prompt' mode much more useful.

The popup doesn't have any animations for the auto-hide timer. It was distracting to see out of the corner of your eye (which is why I had later added a setting) and I didn't want to re-add it with the redesign. The setting disabling animations too is gone.

There's a new 'Bottom Center' positioning. Since I read with my browser on a monitor above my ebook reader it's the easiest position for me to use.

When an error is encountered, the minimized popup shows an error state now (red colors, "Error!" text). If you had the popup auto-hidden previously it wasn't obvious if you weren't signed in.

The build and packaging strategy has changed a bit. Chrome requires manifest version 3 now so some tweaks were needed. I also cleaned up a lot of the code interacting with `chrome`/`browser` by using the webextension-polyfill.