Trim phrases after splitting.
If splitting a phrase results in dangling quote symbols the server will add a different phrase than the extension shows. Like when splitting a phrase that contains a quote with two sentences each sentence might have part of the quotation marks and prompt to add a phrase like `「はい。` but then after trimming when adding to the server you'd get just `はい。` and the extension would consider it a different phrase than `「はい。` so you'd get into a weird state where `「はい。` is still being prompted for but a new `はい。` shows up in the popup. Trimming after splitting makes it so phrases should be identical to whatever the server would do.