Fixed removing cards from phrases other than the current one
This has been a bug for a while that I only just noticed! The `removeCard` method only tried to remove a card from the current phrase. If the card wasn't in the current phrase it'd remove index -1 (the result of the indexOf call).

So if you made one card for each of three phrases, then while sitting on the third phrase you tried to remove the card from the second it'd only remove the card you just made from the third.

I fixed this to now loop through all phrases until it finds the one where the card actually is, instead of always assuming it's in the current phrase.

Added a .projectlaunch.json for projectlaunch.nvim.